Textbook 6.20 Multiples
Write a function isMultiple that determines for a pair of integers whether the second is a multiple of the first. The function should take two integers arguments and return true if the second is a multiple of the first, false otherwise. Use this function in a program that inputs a series of pairs of integers.