Textbook 7.21 Sales Summary
Write a program that will read in 
a) the salesperson number (1 to 4),
b) the product number (1 to 5), and 
c) the total dollar value of that product sold that day
and summarize the total sales by salesperson by product.
All totals should be stored in a two-dimensional array. Print in tabular format with the row being product and column being the salesperson. Include cross totals for each row and column.